pyproject.toml
[project]
name = "ok-serial-relay"
version = "0.1.0"
description = "Serial-port JSON message relay"
readme = "README.md"
authors = [{ name = "Dan Egnor", email = "egnor@ofb.net" }]
requires-python = ">=3.10"
dependencies = [
  "anycrc>=1.3.6",
  "foxglove-sdk>=0.16.1",
  "msgspec>=0.20.0",
  "ok-logging-setup>=0.11",
  "platformdirs>=4.5.1",
  "pyserial>=3.5",
]

[build-system]
requires = ["uv_build>=0.9.13,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = ["black", "mypy", "pytest", "ruff"]

[project.scripts]
okrelay = "ok_serial_relay.okrelay_cli:main"

[tool.black]
line-length = 80

[tool.mypy]
exclude_gitignore = true
files = "."

[tool.uv.build-backend]
module-root = ""
